separated helper comments renderer
